# Fix: Optimize Get Timing Function for animations