This PR properly optimizes the getTimingFunction, which is called during animations, and drops the memize dependency. Applying memoization negatively impacts the JIT's ability to optimize function paths in named easing cases while it helps for custom timings. The suggested change improves performance for both named as well as custom timing cases.

There are 2 distinct cases for `getTimingFunction`: using named easing timings and using custom `cubic-bezier` timings. In the case of named timings the original function proved to be faster then the updated memized function, as the JIT was able to optimize the old switch case fairly well. 

The original performance hit came from dealing with `cubic-bezier`'s due to their required manual string operations and `parseFloat` to convert the string to a `number`. This change caches where needed, storing the `cubic-bezier`s in a key/value object and relies on a `regex` to parse the string. 

Measurements with the original implementation, the memized implementation and the optimized implementation in this PR:

With 1 named value:
```
[Log] original x 4,628 ops/sec ±0.78% (47 runs sampled) (index.ts, line 17)
[Log] memoized x 2,096 ops/sec ±0.50% (47 runs sampled) (index.ts, line 17)
[Log] optimized x 5,666 ops/sec ±0.69% (47 runs sampled) (index.ts, line 17)
```

The optimized is doing about 1.5-2x of the memoized implementation. 
With 1 named value the original implementation is pretty close to the optimized one due to JITs ability to graph the early path out of the switch and apply caching.

With 2 named values:
```
[Log] original x 2,755 ops/sec ±0.68% (45 runs sampled) (index.ts, line 17)
[Log] memoized x 1,410 ops/sec ±1.10% (46 runs sampled) (index.ts, line 17)
[Log] optimized x 4,907 ops/sec ±1.05% (46 runs sampled) (index.ts, line 17)
```

With 2 the difference is still pretty similar, though the original function starts to fall behind.

With 3 named values: 
```
[Log] original x 1,319 ops/sec ±5.07% (39 runs sampled) (index.ts, line 17)
[Log] memoized x 1,264 ops/sec ±0.84% (47 runs sampled) (index.ts, line 17)
[Log] optimized x 4,927 ops/sec ±0.58% (26 runs sampled) (index.ts, line 17)
```

The optimized function is rather unaffected by the named exports and now does 3-4x of both other implementations.
The interesting part here is that the original starts to match the memoized implementation and the hash map is starting to pay off.

However if we run the same test with just 1 unique  `cubic-bezier` it starts to get really interesting.

With 1 cubic-bezier:
```
[Log] original x 176 ops/sec ±4.78% (37 runs sampled) (index.ts, line 17)
[Log] memoized x 2,054 ops/sec ±0.45% (47 runs sampled) (index.ts, line 17)
[Log] optimized x 5,722 ops/sec ±0.69% (46 runs sampled) (index.ts, line 17)
```

Here is where the original implementation drops on the floor, this is due to the string splits and `parseFloat`. Memoize starts to shine because it can cache the response pretty well. 

Though the optimized is still at 2.5x+ over memoized. 

Whether we use 1 or 6 unique  `cubic-bezier`'s, the original remains at the floor:

With 6 unique cubic beziers:
```
[Log] original x 167 ops/sec ±5.40% (38 runs sampled) (index.ts, line 17)
[Log] memoized x 1,113 ops/sec ±2.06% (45 runs sampled) (index.ts, line 17)
[Log] optimized x 4,676 ops/sec ±2.01% (48 runs sampled) (index.ts, line 17)
```

Even the memoized starts to almost drop by 50% - the original function remains almost unchanged. This is the JIT still  being bogged down by the same `parseFloat` bottleneck.

The new implementation still sits pretty stable and now doing 4x+ over memoized. 
If we mix named and cubic, you get this (this was a suggested "real use case") you get the following.

With 2 named and 1 unique cubic bezier:
```
[Log] original x 470 ops/sec ±4.24% (44 runs sampled) (index.ts, line 17)
[Log] memoized x 1,294 ops/sec ±0.78% (47 runs sampled) (index.ts, line 17)
[Log] optimized x 4,844 ops/sec ±1.24% (47 runs sampled) (index.ts, line 17)
```

What's interesting to see is that the original function starts to pick up again, this is because for the 2 named timings the JIT can apply inline caching and apply DFG graphing through the switch case routes. 

The new implementation is rather unchanged, still sitting at a 4x over memoized.

Now a worst case test, all possible named timings and a load of cubic's. 

With 17 named and 30 unique cubic beziers:
```
[Log] original x 246 ops/sec ±7.62% (38 runs sampled) (index.ts, line 17)
[Log] memoized x 245 ops/sec ±1.39% (46 runs sampled) (index.ts, line 17)
[Log] optimized x 3,016 ops/sec ±1.50% (46 runs sampled) (index.ts, line 17)
```

This is where it starts to get funny, both the original and the memoized approach are on the floor. While the optimized function still sits at 3k ops, doing 12x of both. 

Tests run by adding 1000 animation easing requests and measuring performance using `benchmark.js` from jsperf.
Tested on WPEWebKit 2.38 running on a Glass TV. 

_small note_ this will likely even get better once we get FTL and OMG on WPEWebKit in late 2024 😎