@chiefcll 2024-04-16
<img width="678" alt="image" src="https://github.com/lightning-js/renderer/assets/2090941/0fe1136e-2d4a-49f3-9de7-8555a543cc67">

I've updated the test with the new code. Memo is still kicking but when there is one cubic. Let me know if you have the updated test code but I think we still want the memo in there.