@chiefcll 2024-04-16
If you don't want the memo, you just need to cache the results in the timing map rather than returning it right away.