@wouterlucas 2024-04-16
> <img alt="image" width="678" src="https://private-user-images.githubusercontent.com/2090941/322677732-0fe1136e-2d4a-49f3-9de7-8555a543cc67.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MTMyNzc5ODcsIm5iZiI6MTcxMzI3NzY4NywicGF0aCI6Ii8yMDkwOTQxLzMyMjY3NzczMi0wZmUxMTM2ZS0yZDRhLTQ5ZjMtOWRlNy04NTU1YTU0M2NjNjcucG5nP1gtQW16LUFsZ29yaXRobT1BV1M0LUhNQUMtU0hBMjU2JlgtQW16LUNyZWRlbnRpYWw9QUtJQVZDT0RZTFNBNTNQUUs0WkElMkYyMDI0MDQxNiUyRnVzLWVhc3QtMSUyRnMzJTJGYXdzNF9yZXF1ZXN0JlgtQW16LURhdGU9MjAyNDA0MTZUMTQyODA3WiZYLUFtei1FeHBpcmVzPTMwMCZYLUFtei1TaWduYXR1cmU9YmVhNWVkYjQyMjliN2NhNzZkMDA4ZWFjNWRiZTkwMGE2ZWRjNzYxNzQ2MDM5YjM3OWJkNTFlZjBhNTRiNjQzYSZYLUFtei1TaWduZWRIZWFkZXJzPWhvc3QmYWN0b3JfaWQ9MCZrZXlfaWQ9MCZyZXBvX2lkPTAifQ.HSOW8az8bTDQzJp51uzk2mzQbw-t07PTFGhfrSAQ5k8">
> 
> I've updated the test with the new code. Memo is still kicking butt when there is one cubic. Let me know if you have the updated test code but I think we still want the memo in there.

I think you're running an older experiment, prior to caching being added. Feel free to pull the latest!