@chiefcll 2024-04-16
Just to summarize...
1. We're memoizing the cubic-beziers at startup with all the named ones, even though the user may not use any of them. Costing startup time.
2. You've added memoization for each custom cubic-bezier by storing it in the timingMap.

Ideally each cubic-bezier should be calculated and cached when they are needed (which is what the memoize library did). Doing it with the hash map is more efficient though (as its not a generic library doing it). https://www.npmjs.com/package/fast-memoize is supposed to be faster but I didn't have the time to switch that out when I added this. I think we should go back to the original code and just add the timingMap cache.