@wouterlucas 2024-04-16
Going back to the original code will not get you close to the same results. It's even a little faster then memoization the entire function.

To verify this I've ran some tests including: hash map tests, lazy loading of the timings and adding the cache to the original function, here are the results:

3 named + 1 cubic:
```
[Log] original x 366 ops/sec ±6.87% (38 runs sampled) (index.ts, line 21)
[Log] original cached x 1,408 ops/sec ±1.61% (42 runs sampled) (index.ts, line 21)
[Log] memoized x 1,167 ops/sec ±1.74% (45 runs sampled) (index.ts, line 21)
[Log] optimized x 4,806 ops/sec ±0.76% (46 runs sampled) (index.ts, line 21)
[Log] optimized lazy x 4,504 ops/sec ±0.55% (47 runs sampled) (index.ts, line 21)
[Log] hashmap x 2,356 ops/sec ±1.13% (45 runs sampled) (index.ts, line 21)
[Log] hashmap lazy x 2,611 ops/sec ±0.37% (44 runs sampled) (index.ts, line 21)
```

And even though there is a small hit on lazy loading in total throughput, due to the additional IF checks required to make that happen. 

Fully agree that this might make loading/instantiation a lot faster.

For comparison a worst case run, 17 named + 30 cubics:
```
[Log] original x 256 ops/sec ±6.67% (39 runs sampled) (index.ts, line 21)
[Log] original cached x 1,509 ops/sec ±1.94% (44 runs sampled) (index.ts, line 21)
[Log] memoized x 237 ops/sec ±0.98% (45 runs sampled) (index.ts, line 21)
[Log] optimized x 3,021 ops/sec ±0.91% (47 runs sampled) (index.ts, line 21)
[Log] optimized lazy x 2,915 ops/sec ±0.71% (47 runs sampled) (index.ts, line 21)
[Log] hashmap x 2,258 ops/sec ±1.05% (45 runs sampled) (index.ts, line 21)
[Log] hashmap lazy x 2,308 ops/sec ±0.83% (47 runs sampled) (index.ts, line 21)
```

and only 2 named instances:
```
[Log] original x 2,701 ops/sec ±0.74% (46 runs sampled) (index.ts, line 21)
[Log] original cached x 2,497 ops/sec ±0.81% (47 runs sampled) (index.ts, line 21)
[Log] memoized x 1,413 ops/sec ±0.66% (46 runs sampled) (index.ts, line 21)
[Log] optimized x 4,819 ops/sec ±1.06% (47 runs sampled) (index.ts, line 21)
[Log] optimized lazy x 4,518 ops/sec ±0.61% (47 runs sampled) (index.ts, line 21)
[Log] hashmap x 2,519 ops/sec ±0.85% (47 runs sampled) (index.ts, line 21)
[Log] hashmap lazy x 2,848 ops/sec ±1.50% (45 runs sampled) (index.ts, line 21)
```

To summarize:
* No need for an external library, implementing caching where it is needed by ourselves is faster
* Writing optimized JIT-friendly code is fastest
* Lazy loading has a small hit, but can help on load times

Other then that hashmaps seem to be slower then object key/values caching on the box, kind of surprised me to.
Tests are pushed to the experiments repo.

If desired I can push the updated PR with the lazy loading implementation.