@wouterlucas 2024-04-16
Re the hashmaps, definitely something we might want to track. I suspect this might be due to certain optimization layers not being available in 32-bit on WPEWebKit 2.38. If I run the same tests in 64-bit I get a lot different results, there the hashmaps start to really take off and even the lazy loading going 🚀 

Something worthwhile to discuss with the webkit friends, and we should test this again once the new WebKit is made available to us.