@wouterlucas 2024-04-19
> Is creating new functions each time for `linear`, `step-start`, and `step-end` really faster than re-using one copy of the function from a map cache? I'll move this forward cause this has gone on way too long

It is not faster no, but then we go back to the "initial hit" on startup concern that was raised by Chris. This is a sort of middle ground to balance performance vs initial creation hit.

Funny enough it does get faster to lazy load when FTL becomes available in 32-bit. So if you run the experiment on a 64 bit browser you'll see the "lazy" loading approaches get faster. Same as hashmaps, though marginally compared to what we do now.

We could revisit this once those optimizations in the 32-bit browser versions are available to the RDK community.